separation of graph methods
